+++ src/models/fact.rs
@@ -138,7 +138,7 @@ mod tests {
             FactValueType::Integer
         );
         assert_eq!(
-            FactValueType::from(&serde_json::json!(3.14)),
+            FactValueType::from(&serde_json::json!(std::f64::consts::PI)),
             FactValueType::Float
         );
         assert_eq!(